Checking in halfway through the project
Added in the beginnings of RayCast creation. Haven't gotten to detection of hits, just where the laser will route to. About to add in the laser in the next step.

At 20:57 in the video

My headset died, so I can't listen to the information for a bit. Taking a break and will return to the video once it's charged up.